requirements.txt
fastapi
uvicorn[standard]
transformers
torch
bcrypt
python-multipart
python-dotenv
huggingface_hub
